# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory Management API Documentation",
        "description": "Dokumentasi API untuk Sistem Inventaris Tahap 1-4",
        "contact": {
            "email": "admin@inventaris.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server Utama"
        }
    ],
    "paths": {
...
        "/api/suppliers": {
            "get": {
                "tags": [
                    "Partners - Suppliers"
                ],
                "summary": "Ambil daftar supplier beserta kategori komponen yang dipasok",
                "description": "Mendapatkan seluruh daftar supplier/partner pemasok barang beserta kategori komponen yang mereka pasok. Memerlukan hak akses Admin atau Staff.",
                "operationId": "getSupplierList",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data supplier",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Gigabyte Technology"
                                                    },
                                                    "address": {
                                                        "type": "string",
                                                        "example": "New Taipei City, Taiwan"
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "+886-2-8912-4000"
                                                    },
                                                    "category": {
                                                        "type": "string",
                                                        "example": "Motherboard & GPU"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Partners - Suppliers"
                ],
                "summary": "Tambah supplier baru",
                "description": "Menambahkan data supplier baru ke dalam database. Memerlukan hak akses Admin.",
                "operationId": "storeSupplier",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "address",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Gigabyte Technology"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "New Taipei City, Taiwan"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+886-2-8912-4000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Supplier berhasil ditambahkan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Supplier berhasil ditambahkan"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Gigabyte Technology"
                                                },
                                                "address": {
                                                    "type": "string",
                                                    "example": "New Taipei City, Taiwan"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "+886-2-8912-4000"
                                                },
                                                "category": {
                                                    "type": "string",
                                                    "example": "General"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Hanya untuk Admin"
                    },
                    "422": {
                        "description": "Validasi input gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/suppliers/{id}": {
            "put": {
                "tags": [
                    "Partners - Suppliers"
                ],
                "summary": "Update data supplier",
                "description": "Memperbarui data supplier berdasarkan ID. Memerlukan hak akses Admin.",
                "operationId": "updateSupplier",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Supplier",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Gigabyte Tech Co Ltd"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Taipei, Taiwan"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+886-2-8912-4001"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Supplier berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Supplier berhasil diperbarui"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Gigabyte Tech Co Ltd"
                                                },
                                                "address": {
                                                    "type": "string",
                                                    "example": "Taipei, Taiwan"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "+886-2-8912-4001"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Hanya untuk Admin"
                    },
                    "404": {
                        "description": "Supplier tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi update gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Partners - Suppliers"
                ],
                "summary": "Hapus supplier",
                "description": "Menghapus data supplier dari database berdasarkan ID. Memerlukan hak akses Admin.",
                "operationId": "deleteSupplier",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Supplier",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supplier berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Supplier berhasil dihapus"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden - Hanya untuk Admin"
                    },
                    "404": {
                        "description": "Supplier tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}